refactor: split operations.rs into focused modules (part of #777)
Extract special_cases.rs, tags.rs, and s3.rs from operations.rs
to improve code organization. Each module now has a single concern:

- operations.rs: CRUD lifecycle operations (read, create, update, delete)
- special_cases.rs: resource-type-specific attribute handling
- tags.rs: tag conversion between DSL and CloudFormation formats
- s3.rs: S3-specific operations (empty bucket for force_delete)
