+++ carina-provider-awscc/src/provider/operations.rs
@@ -13,7 +13,6 @@ use serde_json::json;
 use super::conversion::{aws_value_to_dsl, dsl_value_to_aws};
 use super::update::build_update_patches;
 use super::{AwsccProvider, get_schema_config};
-use crate::schemas::generated::AwsccSchemaConfig;
 
 impl AwsccProvider {
     /// Read a resource using its configuration
@@ -229,206 +228,5 @@ impl AwsccProvider {
         self.cc_delete_resource(config.aws_type_name, identifier)
             .await
             .map_err(|e| e.for_resource(id.clone()))
-    }
-
-    // =========================================================================
-    // Special Case Handlers
-    // =========================================================================
-
-    /// Handle special attributes that don't follow standard mapping
-    fn read_special_attributes(
-        &self,
-        resource_type: &str,
-        props: &serde_json::Value,
-        attributes: &mut HashMap<String, Value>,
-    ) {
-        match resource_type {
-            "ec2.internet_gateway" => {
-                // Get VPC attachment
-                if let Some(attachments) = props.get("Attachments").and_then(|v| v.as_array())
-                    && let Some(first) = attachments.first()
-                    && let Some(vpc_id) = first.get("VpcId").and_then(|v| v.as_str())
-                {
-                    attributes.insert("vpc_id".to_string(), Value::String(vpc_id.to_string()));
-                }
-            }
-            "ec2.vpc_endpoint" => {
-                // Handle route_table_ids list
-                if let Some(rt_ids) = props.get("RouteTableIds").and_then(|v| v.as_array()) {
-                    let ids: Vec<Value> = rt_ids
-                        .iter()
-                        .filter_map(|v| v.as_str().map(|s| Value::String(s.to_string())))
-                        .collect();
-                    if !ids.is_empty() {
-                        attributes.insert("route_table_ids".to_string(), Value::List(ids));
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    /// Handle special attributes for create
-    fn create_special_attributes(
-        &self,
-        _resource: &Resource,
-        _desired_state: &mut serde_json::Map<String, serde_json::Value>,
-    ) {
-    }
-
-    /// Set default values for create
-    fn set_default_values(
-        &self,
-        resource_type: &str,
-        desired_state: &mut serde_json::Map<String, serde_json::Value>,
-    ) {
-        if resource_type == "ec2.eip" && !desired_state.contains_key("Domain") {
-            desired_state.insert("Domain".to_string(), json!("vpc"));
-        }
-    }
-
-    /// Handle pre-delete operations (e.g., detach IGW from VPC)
-    async fn pre_delete_operations(
-        &self,
-        id: &ResourceId,
-        config: &AwsccSchemaConfig,
-        identifier: &str,
-    ) -> ProviderResult<()> {
-        if id.resource_type == "ec2.internet_gateway" {
-            // Detach from VPC first
-            if let Some(props) = self
-                .cc_get_resource(config.aws_type_name, identifier)
-                .await?
-                && let Some(attachments) = props.get("Attachments").and_then(|v| v.as_array())
-                && !attachments.is_empty()
-            {
-                let patch_ops = vec![json!({"op": "remove", "path": "/Attachments"})];
-                self.cc_update_resource(config.aws_type_name, identifier, patch_ops)
-                    .await
-                    .map_err(|e| {
-                        ProviderError::new(
-                            "Failed to detach Internet Gateway from VPC before deletion",
-                        )
-                        .with_cause(e)
-                        .for_resource(id.clone())
-                    })?;
-            }
-        }
-        Ok(())
-    }
-
-    // =========================================================================
-    // Tag Helpers
-    // =========================================================================
-
-    /// Build tags array for CloudFormation format
-    pub(crate) fn build_tags(&self, user_tags: Option<&Value>) -> Vec<serde_json::Value> {
-        let mut tags = Vec::new();
-        if let Some(Value::Map(user_tags)) = user_tags {
-            for (key, value) in user_tags {
-                if let Value::String(v) = value {
-                    tags.push(json!({"Key": key, "Value": v}));
-                }
-            }
-        }
-        tags
-    }
-
-    /// Parse tags from CloudFormation format to map
-    pub(crate) fn parse_tags(&self, tags_array: &[serde_json::Value]) -> HashMap<String, Value> {
-        let mut tags_map = HashMap::new();
-        for tag in tags_array {
-            if let (Some(key), Some(value)) = (
-                tag.get("Key").and_then(|v| v.as_str()),
-                tag.get("Value").and_then(|v| v.as_str()),
-            ) {
-                tags_map.insert(key.to_string(), Value::String(value.to_string()));
-            }
-        }
-        tags_map
-    }
-
-    /// Create an S3 client from the stored config
-    fn s3_client(&self) -> aws_sdk_s3::Client {
-        aws_sdk_s3::Client::new(&self.aws_config)
-    }
-
-    /// Empty an S3 bucket by deleting all objects and versions
-    async fn empty_s3_bucket(&self, bucket_name: &str) -> ProviderResult<()> {
-        let s3 = self.s3_client();
-
-        // Delete all object versions (handles versioned and non-versioned buckets)
-        let mut key_marker: Option<String> = None;
-        let mut version_id_marker: Option<String> = None;
-
-        loop {
-            let mut req = s3.list_object_versions().bucket(bucket_name).max_keys(1000);
-            if let Some(ref km) = key_marker {
-                req = req.key_marker(km);
-            }
-            if let Some(ref vim) = version_id_marker {
-                req = req.version_id_marker(vim);
-            }
-
-            let response = req
-                .send()
-                .await
-                .map_err(|e| ProviderError::new("Failed to list object versions").with_cause(e))?;
-
-            let mut objects_to_delete = Vec::new();
-
-            // Collect versions
-            for version in response.versions() {
-                if let Some(key) = version.key() {
-                    let mut id = aws_sdk_s3::types::ObjectIdentifier::builder().key(key);
-                    if let Some(vid) = version.version_id() {
-                        id = id.version_id(vid);
-                    }
-                    objects_to_delete.push(id.build().map_err(|e| {
-                        ProviderError::new("Failed to build ObjectIdentifier").with_cause(e)
-                    })?);
-                }
-            }
-
-            // Collect delete markers
-            for marker in response.delete_markers() {
-                if let Some(key) = marker.key() {
-                    let mut id = aws_sdk_s3::types::ObjectIdentifier::builder().key(key);
-                    if let Some(vid) = marker.version_id() {
-                        id = id.version_id(vid);
-                    }
-                    objects_to_delete.push(id.build().map_err(|e| {
-                        ProviderError::new("Failed to build ObjectIdentifier").with_cause(e)
-                    })?);
-                }
-            }
-
-            // Batch delete (max 1000 per request)
-            if !objects_to_delete.is_empty() {
-                let delete = aws_sdk_s3::types::Delete::builder()
-                    .set_objects(Some(objects_to_delete))
-                    .quiet(true)
-                    .build()
-                    .map_err(|e| {
-                        ProviderError::new("Failed to build Delete request").with_cause(e)
-                    })?;
-
-                s3.delete_objects()
-                    .bucket(bucket_name)
-                    .delete(delete)
-                    .send()
-                    .await
-                    .map_err(|e| ProviderError::new("Failed to delete objects").with_cause(e))?;
-            }
-
-            if response.is_truncated() == Some(true) {
-                key_marker = response.next_key_marker().map(|s| s.to_string());
-                version_id_marker = response.next_version_id_marker().map(|s| s.to_string());
-            } else {
-                break;
-            }
-        }
-
-        Ok(())
     }
 }
